yt:git_commit:d2091142c511aa736b907dd35f6f0eef84591abb
yt:last_git_svn_commit:62d1f5cf5428bff587e6f78e6a658d1c25dded57
yt:last_svn_revision:3168903

__FORCE_COMMIT__


git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_2/yt@3170011 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ ytlib/table_client/versioned_block_writer.cpp
@@ -122,7 +122,7 @@ TBlock TSimpleVersionedBlockWriter::FlushBlock()
     auto strings = StringDataStream_.Flush();
     blockParts.insert(blockParts.end(), strings.begin(), strings.end());
 
-    int size = 0;
+    i64 size = 0;
     for (auto& part : blockParts) {
         size += part.Size();
     }
